Fix mouse coords for programs interested with it
In normal usage konsole offsets the position of the mouse by half a char
so that the user can extend the selection to the last column. When using
programs interested with mouse (vim, mc, aptitude, etc...) konsole
shouldn't apply the offset since mouse events should always apply to the
character cell under the pointer. This patch sets the 'edge' flag of
getCharacterPosition to false when _usesMouseTracking is set.